chore: [MR-699] Move Proto Conversions in Replicated State to own Files (#6082)
Protobuf conversions are most of the time not of primary interest when
looking at source code. Splitting them off into separate modules should
improve readability.
