+++ rs/replicated_state/src/canister_state/queues/message_pool/tests.rs
@@ -1,5 +1,7 @@
 use super::*;
+use crate::canister_state::queues::pb_queues;
 use assert_matches::assert_matches;
+use ic_protobuf::proxy::ProxyDecodeError;
 use ic_test_utilities_types::messages::{RequestBuilder, ResponseBuilder};
 use ic_types::messages::{Payload, MAX_INTER_CANISTER_PAYLOAD_IN_BYTES_U64};
 use ic_types::time::UNIX_EPOCH;
