+++ rs/replicated_state/src/canister_state/queues/queue/tests.rs
@@ -1,8 +1,10 @@
 use super::super::message_pool::tests::*;
 use super::super::message_pool::{Class, InboundReference};
 use super::*;
+use crate::canister_state::queues::pb_queues;
 use crate::canister_state::DEFAULT_QUEUE_CAPACITY;
 use assert_matches::assert_matches;
+use ic_protobuf::proxy::ProxyDecodeError;
 use ic_test_utilities_types::ids::{canister_test_id, message_test_id, user_test_id};
 use ic_test_utilities_types::messages::IngressBuilder;
 use proptest::prelude::*;
