+++ rs/replicated_state/src/canister_state/system_state/task_queue.rs
@@ -1,9 +1,9 @@
+pub mod proto;
+
 use crate::ExecutionTask;
 use ic_config::flag_status::FlagStatus;
 use ic_interfaces::execution_environment::ExecutionRoundType;
 use ic_management_canister_types_private::OnLowWasmMemoryHookStatus;
-use ic_protobuf::proxy::ProxyDecodeError;
-use ic_protobuf::state::canister_state_bits::v1 as pb;
 use ic_types::CanisterId;
 use ic_types::NumBytes;
 use num_traits::SaturatingSub;
@@ -248,43 +248,6 @@ impl TaskQueue {
         };
 
         self.paused_or_aborted_task = Some(aborted_task);
-    }
-}
-
-impl From<&TaskQueue> for pb::TaskQueue {
-    fn from(item: &TaskQueue) -> Self {
-        Self {
-            paused_or_aborted_task: item.paused_or_aborted_task.as_ref().map(|task| task.into()),
-            on_low_wasm_memory_hook_status: pb::OnLowWasmMemoryHookStatus::from(
-                &item.on_low_wasm_memory_hook_status,
-            )
-            .into(),
-            queue: item.queue.iter().map(|task| task.into()).collect(),
-        }
-    }
-}
-
-impl TryFrom<pb::TaskQueue> for TaskQueue {
-    type Error = ProxyDecodeError;
-
-    fn try_from(item: pb::TaskQueue) -> Result<Self, Self::Error> {
-        Ok(Self {
-            paused_or_aborted_task: item
-                .paused_or_aborted_task
-                .map(|task| task.try_into())
-                .transpose()?,
-            on_low_wasm_memory_hook_status: pb::OnLowWasmMemoryHookStatus::try_from(
-                item.on_low_wasm_memory_hook_status,
-            )
-            .map_err(|e| ProxyDecodeError::Other(
-                format!("Error while trying to decode pb::TaskQueue::on_low_wasm_memory_hook_status, {:?}", e)))?
-            .try_into()?,
-            queue: item
-                .queue
-                .into_iter()
-                .map(|task| task.try_into())
-                .collect::<Result<VecDeque<_>, _>>()?,
-        })
     }
 }
