+++ rs/replicated_state/src/canister_state/system_state/wasm_chunk_store.rs
@@ -1,12 +1,13 @@
 use std::{collections::BTreeMap, sync::Arc};
 
-use ic_protobuf::{proxy::ProxyDecodeError, state::canister_state_bits::v1 as pb};
 use ic_sys::{PageBytes, PageIndex, PAGE_SIZE};
 use ic_types::{NumBytes, NumOsPages};
 use ic_validate_eq::ValidateEq;
 use ic_validate_eq_derive::ValidateEq;
 
 use crate::{page_map::PageAllocatorFileDescriptor, PageMap};
+
+pub mod proto;
 
 /// This is the _maximum_ chunk size. A chunk may take up as little space as
 /// a single OS page. However, the cycles cost of maintaining a chunk in the
@@ -221,50 +222,6 @@ pub struct WasmChunkStoreMetadata {
     chunks: BTreeMap<WasmChunkHash, ChunkInfo>,
     /// Total size of the data in the chunk store.
     size: NumOsPages,
-}
-
-impl From<&WasmChunkStoreMetadata> for pb::WasmChunkStoreMetadata {
-    fn from(item: &WasmChunkStoreMetadata) -> Self {
-        let chunks = item
-            .chunks
-            .iter()
-            .map(|(hash, ChunkInfo { index, length })| pb::WasmChunkData {
-                hash: hash.to_vec(),
-                index: *index,
-                length: *length,
-            })
-            .collect::<Vec<_>>();
-        let size = item.size.get();
-        pb::WasmChunkStoreMetadata { chunks, size }
-    }
-}
-
-impl TryFrom<pb::WasmChunkStoreMetadata> for WasmChunkStoreMetadata {
-    type Error = ProxyDecodeError;
-
-    fn try_from(value: pb::WasmChunkStoreMetadata) -> Result<Self, Self::Error> {
-        let mut chunks = BTreeMap::new();
-        for chunk in value.chunks {
-            let hash: [u8; 32] =
-                chunk
-                    .hash
-                    .try_into()
-                    .map_err(|e| ProxyDecodeError::ValueOutOfRange {
-                        typ: "[u8; 32]",
-                        err: format!("Failed to convert vector to fixed size arrary: {:?}", e),
-                    })?;
-            chunks.insert(
-                hash,
-                ChunkInfo {
-                    index: chunk.index,
-                    length: chunk.length,
-                },
-            );
-        }
-
-        let size = value.size.into();
-        Ok(Self { chunks, size })
-    }
 }
 
 #[cfg(test)]
